requirements.txt
requests
urllib3
ollama
jira
python-dotenv